refactor: :recycle: 20250305 ダイクストラ部分のモジュール化
