+++ v1/roulette.html
@@ -207,7 +207,6 @@
     <!--Panzoom-->
     <script src="https://unpkg.com/@panzoom/panzoom@4.5.1/dist/panzoom.min.js"></script>
     <!--JavaScript-->
-    <script type="module" src="../js/rouletteLocation.js" charset="UTF-8" defer="defer"></script>
     <script type="module" src="../js/roulette.js" charset="UTF-8" defer="defer"></script>
 </body>
 <footer>
